requirements.txt
google-cloud-storage
python-dotenv
pandas
requests
google-cloud-bigquery
pyarrow
db-dtypes
pandas-gbq
pymongo
streamlit
bcrypt